Commit of second version
Commit of revised version after first set of reviewer's comments.

+++ project-system-monitor/include/process.h
@@ -12,7 +12,7 @@ class Process {
         int Pid() const;                          // DONE - TODO: See src/process.cpp
         string User() const;                      // DONE - TODO: See src/process.cpp
         string Command() const;                   // DONE - TODO: See src/process.cpp
-        float CpuUtilization() const;             // DONE - TODO: See src/process.cpp
+        double CpuUtilization() const;             // DONE - TODO: See src/process.cpp
         string Ram() const;                       // DONE - TODO: See src/process.cpp
         long int UpTime() const;                  // DONE - TODO: See src/process.cpp
         bool operator<(Process const& a) const;   // DONE - TODO: See src/process.cpp
