+++ project-system-monitor/include/system.h
@@ -9,7 +9,7 @@ class System {
     public:
         Processor& Cpu();                   // DONE - TODO: See src/system.cpp
         vector<Process>& Processes();       // DONE - TODO: See src/system.cpp
-        float MemoryUtilization();          // DONE - TODO: See src/system.cpp
+        double MemoryUtilization();          // DONE - TODO: See src/system.cpp
         long UpTime();                      // DONE - TODO: See src/system.cpp
         int TotalProcesses();               // DONE - TODO: See src/system.cpp
         int RunningProcesses();             // DONE - TODO: See src/system.cpp
